requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
joblib>=1.3.0
MetaTrader5>=5.0.0
python-dotenv>=1.0.0
firebase-admin>=6.5.0